Avoid getting version from spec when not DKMS build
In the Makefile, MOD_VERSION is only used to set DKMS_SRC_DIR which is
only used in the DKMS build. In that case, the tarball includes the spec
file. However, the build script does not export the spec file in the
build tarball.

There is no need to get MOD_VERSION in any build except when calling
DKMS. Therefore no point in generating it on every run of make. ':='
is intended to optimize performence but here it causes useless
invocations.

Signed-off-by: Tzafrir Cohen <nvidia@cohens.org.il>